requirements.txt
absl-py==0.11.0
astunparse==1.6.3
cachetools==4.2.0
certifi==2020.12.5
chardet==4.0.0
cmake==3.18.4.post1
cycler==0.10.0
decorator==4.4.2
dlib==19.19.0
efficientnet==1.0.0
flatbuffers==1.12
gast==0.3.3
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
image-classifiers==1.0.0
imageio==2.9.0
importlib-metadata==3.3.0
imutils==0.5.3
Keras==2.4.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.3
matplotlib==3.3.3
networkx==2.5
numpy==1.19.4
oauthlib==3.1.0
opencv-python==3.4.3.18
opt-einsum==3.3.0
Pillow==8.0.1
protobuf==3.14.0
py4j==0.10.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
pyspark==3.0.1
python-dateutil==2.8.1
PyWavelets==1.1.1
PyYAML==5.3.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.6
scikit-image==0.17.2
scipy==1.5.4
segmentation-models==1.0.1
six==1.15.0
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow-estimator==2.4.0
tensorflow-gpu==2.4.0
termcolor==1.1.0
tifffile==2020.9.3
typing-extensions==3.7.4.3
urllib3==1.26.2
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.0
